requirements.txt
streamlit>=1.26.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
